Move wgsl into own file
Part of #221145
